requirements.txt
aiogram==3.4.1
google-generativeai
python-dotenv